# src/configs/swagger/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Vey Badminton WS API",
    "description": "API documentation for Vey Badminton WS"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "tags": [
    {
      "name": "API",
      "description": "API"
    },
    {
      "name": "User",
      "description": "User package"
    },
    {
      "name": "Auth",
      "description": "Auth package"
    },
    {
      "name": "Product",
      "description": "Product package"
    },
    {
      "name": "Order",
      "description": "Order package"
    },
    {
      "name": "Customer",
      "description": "Customer package"
    }
  ],
  "schemes": [
    "http",
    "https"
  ],
  "securityDefinitions": {
    "bearerAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "For accessing the API a valid JWT token must be passed in all the queries in the 'Authorization' header. A valid JWT token is generated by the API and retured as answer of a call to the route /login giving a valid user & password. The following syntax must be used in the 'Authorization' header : Bearer: your_jwt_token."
    }
  },
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/users/": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/users/{id}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/users/create": {
      "post": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/users/{id}/edit": {
      "put": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/users/{id}/delete": {
      "delete": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "definitions": {
    "User": {
      "type": "object",
      "properties": {
        "type": {
          "type": "string",
          "example": "object"
        },
        "properties": {
          "type": "object",
          "properties": {
            "user_id": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "integer"
                },
                "description": {
                  "type": "string",
                  "example": "ID của người dùng"
                }
              }
            },
            "user_name": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "string"
                },
                "description": {
                  "type": "string",
                  "example": "Tên người dùng"
                }
              }
            },
            "user_email": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "string"
                },
                "description": {
                  "type": "string",
                  "example": "Email của người dùng"
                },
                "nullable": {
                  "type": "boolean",
                  "example": true
                }
              }
            },
            "user_avatar": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "string"
                },
                "description": {
                  "type": "string",
                  "example": "Đường dẫn đến ảnh đại diện của người dùng"
                }
              }
            },
            "user_phone": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "string"
                },
                "description": {
                  "type": "string",
                  "example": "Số điện thoại của người dùng"
                }
              }
            },
            "user_full_address": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "string"
                },
                "description": {
                  "type": "string",
                  "example": "Địa chỉ đầy đủ của người dùng"
                },
                "nullable": {
                  "type": "boolean",
                  "example": true
                }
              }
            },
            "user_detail_address": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "string"
                },
                "description": {
                  "type": "string",
                  "example": "Địa chỉ chi tiết của người dùng"
                },
                "nullable": {
                  "type": "boolean",
                  "example": true
                }
              }
            },
            "user_provice_id": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "string"
                },
                "description": {
                  "type": "string",
                  "example": "ID tỉnh/thành phố của người dùng"
                },
                "nullable": {
                  "type": "boolean",
                  "example": true
                }
              }
            },
            "user_district_id": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "string"
                },
                "description": {
                  "type": "string",
                  "example": "ID quận/huyện của người dùng"
                },
                "nullable": {
                  "type": "boolean",
                  "example": true
                }
              }
            },
            "user_ward_id": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "string"
                },
                "description": {
                  "type": "string",
                  "example": "ID phường/xã của người dùng"
                },
                "nullable": {
                  "type": "boolean",
                  "example": true
                }
              }
            },
            "user_code": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "string"
                },
                "description": {
                  "type": "string",
                  "example": "Mã người dùng"
                }
              }
            },
            "user_status": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "integer"
                },
                "description": {
                  "type": "string",
                  "example": "Trạng thái của người dùng"
                },
                "nullable": {
                  "type": "boolean",
                  "example": true
                }
              }
            },
            "user_birthday": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "integer"
                },
                "format": {
                  "type": "string",
                  "example": "int64"
                },
                "description": {
                  "type": "string",
                  "example": "Ngày sinh của người dùng (timestamp)"
                },
                "nullable": {
                  "type": "boolean",
                  "example": true
                }
              }
            },
            "user_description": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "string"
                },
                "description": {
                  "type": "string",
                  "example": "Mô tả về người dùng"
                },
                "nullable": {
                  "type": "boolean",
                  "example": true
                }
              }
            },
            "user_properties": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "string"
                },
                "description": {
                  "type": "string",
                  "example": "Các thuộc tính khác của người dùng"
                },
                "nullable": {
                  "type": "boolean",
                  "example": true
                }
              }
            },
            "user_config": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "integer"
                },
                "description": {
                  "type": "string",
                  "example": "Cấu hình của người dùng"
                }
              }
            },
            "user_password": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "string"
                },
                "description": {
                  "type": "string",
                  "example": "Mật khẩu của người dùng (đã được băm)"
                }
              }
            },
            "user_role_id": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "integer"
                },
                "description": {
                  "type": "string",
                  "example": "ID vai trò của người dùng"
                },
                "nullable": {
                  "type": "boolean",
                  "example": true
                }
              }
            },
            "user_department_id": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "integer"
                },
                "description": {
                  "type": "string",
                  "example": "ID phòng ban của người dùng"
                },
                "nullable": {
                  "type": "boolean",
                  "example": true
                }
              }
            },
            "user_position_id": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "integer"
                },
                "description": {
                  "type": "string",
                  "example": "ID vị trí của người dùng"
                },
                "nullable": {
                  "type": "boolean",
                  "example": true
                }
              }
            },
            "user_create_by": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "integer"
                },
                "description": {
                  "type": "string",
                  "example": "ID người tạo người dùng"
                },
                "nullable": {
                  "type": "boolean",
                  "example": true
                }
              }
            },
            "user_join_date": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "integer"
                },
                "format": {
                  "type": "string",
                  "example": "int64"
                },
                "description": {
                  "type": "string",
                  "example": "Ngày tham gia của người dùng (timestamp)"
                },
                "nullable": {
                  "type": "boolean",
                  "example": true
                }
              }
            },
            "user_deleted_at": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "integer"
                },
                "format": {
                  "type": "string",
                  "example": "int64"
                },
                "description": {
                  "type": "string",
                  "example": "Thời điểm người dùng bị xóa (timestamp)"
                },
                "nullable": {
                  "type": "boolean",
                  "example": true
                }
              }
            },
            "user_created_at": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "integer"
                },
                "format": {
                  "type": "string",
                  "example": "int64"
                },
                "description": {
                  "type": "string",
                  "example": "Thời điểm người dùng được tạo (timestamp)"
                },
                "nullable": {
                  "type": "boolean",
                  "example": true
                }
              }
            },
            "user_updated_at": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "integer"
                },
                "format": {
                  "type": "string",
                  "example": "int64"
                },
                "description": {
                  "type": "string",
                  "example": "Thời điểm người dùng được cập nhật lần cuối (timestamp)"
                },
                "nullable": {
                  "type": "boolean",
                  "example": true
                }
              }
            }
          }
        }
      }
    }
  }
}